requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
google-cloud-storage==2.14.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4 # For Password Hashing
openpyxl==3.1.2
httpx==0.27.0
alembic
python-dotenv
